Added new property Profile in User class

+++ Automarket.Domain/Entity/Profile.cs
@@ -5,7 +5,7 @@ namespace Automarket.Domain.Entity
 	{
 		public long Id { get; set; }
 
-		public byte Age { get; set; }
+		public short Age { get; set; }
 
 		public string? Address { get; set; }
